refactor: split out middleware into own package and add test scaffold
