+++ authnz-service-proxy/middleware/middleware_test.go
@@ -1,0 +1,26 @@
+package middleware
+
+import "testing"
+
+func TestGetTargetConfig(t *testing.T) {
+
+}
+
+func TestVerify(t *testing.T) {
+
+}
+
+func TestAuthenticate(t *testing.T) {
+
+}
+
+func TestIsAuthenticationValid(t *testing.T) {
+
+}
+
+func TestRandString(t *testing.T) {
+
+}
+
+func TestToCookie(t *testing.T) {
+}
